requirements.txt
fastapi
uvicorn
mysql-connector-python
selenium
webdriver-manager
